Incorporating Fengming's commits 844f5044 and b65c9f88.

+++ src/runmodule/Integrator.h
@@ -123,7 +123,7 @@ class Integrator{
        int blackhol;
        float inittol;
        int maxit;
-       long maxitmon;
+       int maxitmon;
        int retry;
 
        int syint;
